+++ tests/Silex/Tests/Provider/TwigServiceProviderTest.php
@@ -84,13 +84,13 @@ class TwigServiceProviderTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('/foo.css?1', $app['twig']->render('hello'));
     }
 
-    public function testAppVariable()
+    public function testGlobalVariable()
     {
         $app = new Application();
         $app['request_stack']->push(Request::create('/?name=Fabien'));
 
         $app->register(new TwigServiceProvider(), array(
-            'twig.templates' => array('hello' => '{{ app.request.get("name") }}'),
+            'twig.templates' => array('hello' => '{{ global.request.get("name") }}'),
         ));
 
         $this->assertEquals('Fabien', $app['twig']->render('hello'));